libarchive 2.2.3   * "compression_program" support uses an external program   * Portability: no longer uses "struct stat" as a primary     data interchange structure internally   * Part of the above: refactor archive_entry to separate     out copy_stat() and stat() functions   * More complete tests for archive_entry   * Finish archive_entry_clone()   * Isolate major()/minor()/makedev() in archive_entry; remove     these from everywhere else.   * Bug fix: properly handle decompression look-ahead at end-of-data   * Bug fixes to 'ar' support   * Fix memory leak in ZIP reader   * Portability: better timegm() emulation in iso9660 reader   * New write_disk flags to suppress auto dir creation and not     overwrite newer files (for future cpio front-end)   * Simplify trailing-'/' fixup when writing tar and pax   * Test enhancements:  fix various compiler warnings, improve     portability, add lots of new tests.   * Documentation: document new functions, first draft of     libarchive_internals.3
MFC after: 14 days
Thanks to: Joerg Sonnenberger (compression_program)
Thanks to: Kai Wang (ar)
Thanks to: Colin Percival (many small fixes)
Thanks to: Many others who sent me various patches and problem reports.

+++ lib/libarchive/test/test_read_position.c
@@ -48,7 +48,7 @@ DEFINE_TEST(test_read_position)
 	assertA(0 == archive_write_open_memory(a, buff, sizeof(buff), &write_pos));
 	assertA(0 == archive_write_header(a, ae));
 	archive_entry_free(ae);
-	assertA(data_size == archive_write_data(a, nulls, sizeof(nulls)));
+	assertA(data_size == (size_t)archive_write_data(a, nulls, sizeof(nulls)));
 #if ARCHIVE_API_VERSION > 1
 	assertA(0 == archive_write_finish(a));
 #else
